Removed unecess Async comp
